gcp/requirements.txt
functions-framework==3.*
tensorflow-cpu==2.16.1
pillow>=10.0.0
google-cloud-storage>=2.0.0
numpy>=1.24.0